chore(j-s): Update substances and combine drugs and prescribed medication lists
