+++ libs/judicial-system/types/src/lib/indictmentCount.ts
@@ -1,51 +1,13 @@
+import {
+  ILLEGAL_DRUGS_AND_PRESCRIPTION_DRUGS_DRIVING,
+  Substance,
+} from './substances'
+
 export enum IndictmentCountOffense {
   DRIVING_WITHOUT_LICENCE = 'DRIVING_WITHOUT_LICENCE',
   DRUNK_DRIVING = 'DRUNK_DRIVING',
   ILLEGAL_DRUGS_DRIVING = 'ILLEGAL_DRUGS_DRIVING',
   PRESCRIPTION_DRUGS_DRIVING = 'PRESCRIPTION_DRUGS_DRIVING',
-}
-
-export enum Substance {
-  ALCOHOL = 'ALCOHOL',
-  AMPHETAMINE = 'AMPHETAMINE',
-  ETIZOLAM = 'ETIZOLAM',
-  PHENAZEPAM = 'PHENAZEPAM',
-  KETAMINE = 'KETAMINE',
-  KETOBEMIDONE = 'KETOBEMIDONE',
-  COCAINE = 'COCAINE',
-  MDMA = 'MDMA',
-  METHAMPHETAMINE = 'METHAMPHETAMINE',
-  METHYLPHENIDATE = 'METHYLPHENIDATE',
-  O_DESMETHYLTRAMADOL = 'O_DESMETHYLTRAMADOL',
-  TETRAHYDROCANNABINOL = 'TETRAHYDROCANNABINOL',
-  TRAMADOL = 'TRAMADOL',
-  ZOPICLONE = 'ZOPICLONE',
-  ALPRAZOLAM = 'ALPRAZOLAM',
-  BROMAZEPAM = 'BROMAZEPAM',
-  BUPRENORPHINE = 'BUPRENORPHINE',
-  DEMOXEAPAM = 'DEMOXEAPAM',
-  DESMETHYLCHLORDIAZEPOXIDE = 'DESMETHYLCHLORDIAZEPOXIDE',
-  DIAZEPAM = 'DIAZEPAM',
-  FENTANYL = 'FENTANYL',
-  FLUNITRAZEPAM = 'FLUNITRAZEPAM',
-  GABAPENTIN = 'GABAPENTIN',
-  CLONAZEPAM = 'CLONAZEPAM',
-  CHLORDIAZEPOXIDE = 'CHLORDIAZEPOXIDE',
-  CODEINE = 'CODEINE',
-  LORAZEPAM = 'LORAZEPAM',
-  METHADONE = 'METHADONE',
-  MIDAZOLAM = 'MIDAZOLAM',
-  MORPHINE = 'MORPHINE',
-  NITRAZEPAM = 'NITRAZEPAM',
-  NORBUPRENORPHINE = 'NORBUPRENORPHINE',
-  NORDIAZEPAM = 'NORDIAZEPAM',
-  OXAZEPAM = 'OXAZEPAM',
-  OXYCODONE = 'OXYCODONE',
-  PREGABALIN = 'PREGABALIN',
-  QUETIAPINE = 'QUETIAPINE',
-  TEMAZEPAM = 'TEMAZEPAM',
-  TRIAZOLAM = 'TRIAZOLAM',
-  ZOLPIDEM = 'ZOLPIDEM',
 }
 
 export type SubstanceMap = { [key in Substance]?: string }
@@ -55,47 +17,8 @@ export const offenseSubstances: {
 } = {
   [IndictmentCountOffense.DRIVING_WITHOUT_LICENCE]: [],
   [IndictmentCountOffense.DRUNK_DRIVING]: [Substance.ALCOHOL],
-  [IndictmentCountOffense.ILLEGAL_DRUGS_DRIVING]: [
-    Substance.AMPHETAMINE,
-    Substance.ETIZOLAM,
-    Substance.PHENAZEPAM,
-    Substance.KETAMINE,
-    Substance.KETOBEMIDONE,
-    Substance.COCAINE,
-    Substance.MDMA,
-    Substance.METHAMPHETAMINE,
-    Substance.METHYLPHENIDATE,
-    Substance.O_DESMETHYLTRAMADOL,
-    Substance.TETRAHYDROCANNABINOL,
-    Substance.TRAMADOL,
-    Substance.ZOPICLONE,
-  ],
-  [IndictmentCountOffense.PRESCRIPTION_DRUGS_DRIVING]: [
-    Substance.ALPRAZOLAM,
-    Substance.BROMAZEPAM,
-    Substance.BUPRENORPHINE,
-    Substance.DEMOXEAPAM,
-    Substance.DESMETHYLCHLORDIAZEPOXIDE,
-    Substance.DIAZEPAM,
-    Substance.FENTANYL,
-    Substance.FLUNITRAZEPAM,
-    Substance.GABAPENTIN,
-    Substance.CLONAZEPAM,
-    Substance.CHLORDIAZEPOXIDE,
-    Substance.CODEINE,
-    Substance.LORAZEPAM,
-    Substance.METHADONE,
-    Substance.MIDAZOLAM,
-    Substance.MORPHINE,
-    Substance.NITRAZEPAM,
-    Substance.NORBUPRENORPHINE,
-    Substance.NORDIAZEPAM,
-    Substance.OXAZEPAM,
-    Substance.OXYCODONE,
-    Substance.PREGABALIN,
-    Substance.QUETIAPINE,
-    Substance.TEMAZEPAM,
-    Substance.TRIAZOLAM,
-    Substance.ZOLPIDEM,
-  ],
+  [IndictmentCountOffense.ILLEGAL_DRUGS_DRIVING]:
+    ILLEGAL_DRUGS_AND_PRESCRIPTION_DRUGS_DRIVING,
+  [IndictmentCountOffense.PRESCRIPTION_DRUGS_DRIVING]:
+    ILLEGAL_DRUGS_AND_PRESCRIPTION_DRUGS_DRIVING,
 }
