+++ libs/judicial-system/types/src/lib/substances.ts
@@ -1,0 +1,88 @@
+export enum Substance {
+  ALCOHOL = 'ALCOHOL',
+  ALPRAZOLAM = 'ALPRAZOLAM',
+  AMPHETAMINE = 'AMPHETAMINE',
+  BROMAZEPAM = 'BROMAZEPAM',
+  BROMAZOLAM = 'BROMAZOLAM',
+  BUPRENORPHINE = 'BUPRENORPHINE',
+  CHLORDIAZEPOXIDE = 'CHLORDIAZEPOXIDE',
+  CLONAZEPAM = 'CLONAZEPAM',
+  COCAINE = 'COCAINE',
+  CODEINE = 'CODEINE',
+  DEMOXEAPAM = 'DEMOXEAPAM',
+  DESMETHYLCHLORDIAZEPOXIDE = 'DESMETHYLCHLORDIAZEPOXIDE',
+  DIAZEPAM = 'DIAZEPAM',
+  ETIZOLAM = 'ETIZOLAM',
+  FENTANYL = 'FENTANYL',
+  FLUALPRAZOLAM = 'FLUALPRAZOLAM',
+  FLUNITRAZEPAM = 'FLUNITRAZEPAM',
+  GABAPENTIN = 'GABAPENTIN',
+  KETAMINE = 'KETAMINE',
+  KETOBEMIDONE = 'KETOBEMIDONE',
+  LORAZEPAM = 'LORAZEPAM',
+  MDMA = 'MDMA',
+  METHADONE = 'METHADONE',
+  METHAMPHETAMINE = 'METHAMPHETAMINE',
+  METHYLPHENIDATE = 'METHYLPHENIDATE',
+  MIDAZOLAM = 'MIDAZOLAM',
+  MORPHINE = 'MORPHINE',
+  NITRAZEPAM = 'NITRAZEPAM',
+  NORBUPRENORPHINE = 'NORBUPRENORPHINE',
+  NORDIAZEPAM = 'NORDIAZEPAM',
+  O_DESMETHYLTRAMADOL = 'O_DESMETHYLTRAMADOL',
+  OXAZEPAM = 'OXAZEPAM',
+  OXYCODONE = 'OXYCODONE',
+  PHENAZEPAM = 'PHENAZEPAM',
+  PREGABALIN = 'PREGABALIN',
+  QUETIAPINE = 'QUETIAPINE',
+  TEMAZEPAM = 'TEMAZEPAM',
+  TETRAHYDROCANNABINOL = 'TETRAHYDROCANNABINOL',
+  TRAMADOL = 'TRAMADOL',
+  TRIAZOLAM = 'TRIAZOLAM',
+  ZOLPIDEM = 'ZOLPIDEM',
+  ZOPICLONE = 'ZOPICLONE',
+}
+
+export const ILLEGAL_DRUGS_AND_PRESCRIPTION_DRUGS_DRIVING = [
+  Substance.AMPHETAMINE,
+  Substance.ALPRAZOLAM,
+  Substance.BROMAZEPAM,
+  Substance.BROMAZOLAM,
+  Substance.BUPRENORPHINE,
+  Substance.CHLORDIAZEPOXIDE,
+  Substance.CLONAZEPAM,
+  Substance.COCAINE,
+  Substance.CODEINE,
+  Substance.DEMOXEAPAM,
+  Substance.DESMETHYLCHLORDIAZEPOXIDE,
+  Substance.DIAZEPAM,
+  Substance.ETIZOLAM,
+  Substance.FENTANYL,
+  Substance.FLUALPRAZOLAM,
+  Substance.FLUNITRAZEPAM,
+  Substance.GABAPENTIN,
+  Substance.KETAMINE,
+  Substance.KETOBEMIDONE,
+  Substance.LORAZEPAM,
+  Substance.MDMA,
+  Substance.METHADONE,
+  Substance.METHAMPHETAMINE,
+  Substance.METHYLPHENIDATE,
+  Substance.MIDAZOLAM,
+  Substance.MORPHINE,
+  Substance.NITRAZEPAM,
+  Substance.NORBUPRENORPHINE,
+  Substance.NORDIAZEPAM,
+  Substance.O_DESMETHYLTRAMADOL,
+  Substance.OXAZEPAM,
+  Substance.OXYCODONE,
+  Substance.PHENAZEPAM,
+  Substance.PREGABALIN,
+  Substance.QUETIAPINE,
+  Substance.TEMAZEPAM,
+  Substance.TETRAHYDROCANNABINOL,
+  Substance.TRAMADOL,
+  Substance.TRIAZOLAM,
+  Substance.ZOLPIDEM,
+  Substance.ZOPICLONE,
+]
